fix: common project should not depend on components project
